Extract Stripe billing from functions index.js into stripe.js
index.js had grown to ~800 lines, mostly inline Stripe billing while
everything else lives in modules it re-exports. Now it follows the same
pattern for billing:

- public/functions/stripe.js (new): createStripeSession,
  checkActiveSubscriptions, createStripeBillingPortal,
  handleSubscriptionWebhook, stripeWebhook, plus the private helpers
  (markCheckoutSessionStatus, handleInvoicePaymentFailed)
- grantPurchasedTokens moves to token-management.js beside its sibling
  chargeGenerationTokens — all tokenProfile mutations + tokenLog writes
  in one module (token-packs.js stays firebase-free so the mocha test
  can require it directly)
- index.js re-exports the five Stripe functions and drops the imports
  only the moved code used; down from 795 to 191 lines

Code moved verbatim — no behavior change. Function names are unchanged
so deployment targets stay the same. pricing-sync.test.js now reads
PRICE_CONFIG from stripe.js.
